backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
pandas
pyjwt
requests
svix
cryptography
gunicorn
